Update README.md
Despite failure to understand lines and hashtags, I am committing to these changes.   Well now I'm in a pickle. This thing thinks I'm in the code of conduct branch, but I was in the edit readme branch. Great.  I'll never be a super programmer now.  I think I'll create a new branch for this.