Rays Journal of big ideas
Whoaa!  I was typing a bunch of stuff up there after "Rays Journal of big ideas"  e.g.  "to be pursued if they can be remembered", when suddenly the system told me to use LESS than 50 characters in order to have a Great Commit!   As a future super-developer, Ray wants to have a GREAT commit.  So here it is, made directly to the master branch.